Re-extended certain classes to maximize polymorphism. The current class inheritance is as follows:
   	GenericUser --> User --> Worker --> Manager
	     |
	     ---------> Administrator

	   *All classes extend each other starting from User except the
	   Administrator class. Since it cant do what the User does it
	   only extends the GenericUser class

Renamed the "OverallUser" class to "GenericUser" for clarity

Also added blank dummy methods that will be implemented later. These
methods detail what each User Type will be able to do in the future. Ps:
if you think they can be named better rename them.

Also added "date" local variable to the GenericUser class which is
automatically filled with that day's date and time of which the User object was
created.